Refactoring. Create new class WeatherData to hold downloadWeatherData logic.
