+++ src/main/java/com/retrievedata/AccuweatherLocationHandler.java
@@ -1,9 +1,5 @@
-package com.dataretrieval;
+package com.retrievedata;
 
-import org.apache.http.HttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
 import org.json.JSONArray;
 import org.json.JSONObject;
