+++ src/main/java/com/retrievedata/LocationDetails.java
@@ -1,4 +1,4 @@
-package com.dataretrieval;
+package com.retrievedata;
 
 import lombok.Getter;
 import lombok.Setter;
