+++ src/main/java/com/retrievedata/WeatherData.java
@@ -1,0 +1,35 @@
+package com.retrievedata;
+
+import org.apache.http.HttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.json.JSONObject;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+
+public class WeatherData {
+
+    public static JSONObject downloadWeatherData(String url) throws IOException {
+
+        CloseableHttpClient httpClient = HttpClientBuilder.create().build();
+        HttpGet getRequest = new HttpGet(url);
+        HttpResponse response = httpClient.execute(getRequest);
+
+        if (response.getStatusLine().getStatusCode() == 200) {
+            BufferedReader reader = new BufferedReader(new InputStreamReader(response.getEntity().getContent()));
+            StringBuilder result = new StringBuilder();
+            String line;
+            while ((line = reader.readLine()) != null) {
+                result.append(line);
+            }
+            return new JSONObject(result.toString());
+        } else {
+            System.err.println("Failed to retrieve data. HTTP Error Code: " + response.getStatusLine().getStatusCode());
+            return null;
+        }
+    }
+}
+
