+++ src/main/java/com/retrievedata/openweather/Main.java
@@ -1,4 +1,4 @@
-package com.dataretrieval.openweather;
+package com.retrievedata.openweather;
 
 import com.entities.WeatherOpenweather;
 import org.json.JSONObject;
